requirementsM2.txt
arxiv
pdfplumber
requests
python-dotenv
qdrant-client
